Split epnf→fluent AST munge step into separate module
